+++ package.json
@@ -29,7 +29,7 @@
     "date-fns": "2.28.0",
     "enquirer": "^2.3.6",
     "next": "^12.1.5",
-    "react": "^18.0.0",
+    "react": "^18.1.0",
     "react-dom": "^18.0.0",
     "react-select": "^5.3.0",
     "slugify": "1.6.5"
